♻️ refactor(core): functions using kqueue() splited to independent file to prepare linux portability Also refactored the testing website
