+++ core/handle_connections.cpp
@@ -48,46 +48,16 @@ int	accept_new_client(const int kq, const int event_fd,
 	return 0;
 }
 
-//Allocate a buffer with the size available into the buffer from the socket
-//connection, receive it and append it to the client requests string
-int	read_request(const struct kevent& event, Client& client)
+//Sets all our IP:PORT listening sockets in the kqueue so they are monitored
+//for read event
+int	monitor_network_sockets(const int kq,
+				const std::map<int, std::pair<std::string, int> > & server_map)
 {
-	char	*buf;
-	ssize_t	len, buflen;
-
-	__D_DISPLAY("data flag from read event: [" << event.data << "]");
-	buflen = event.data;
-	buf = new char [buflen];
-	len = recv(event.ident, buf, buflen, 0);
-	__D_DISPLAY_RECV(event.ident, len);
-	if (len > 0)
-		client.setRequest(buf, len);
-	delete [] buf;
-	return len;
-}
-
-//Send response & set write event to the client again if all response couldn't
-//be sent. Removes data that has been sent from the response buffer
-int	send_response(const int kq, const struct kevent& event, Client& client)
-{
-	ssize_t	len;
-
-	//event->data contains space remaining in the write buffer
-	__D_DISPLAY("data flag from write event: " << event.data);
-	if (event.data < static_cast<long>(client.getLenToSend()))
+	for (std::map<int, std::pair<std::string, int> >::const_iterator
+			i = server_map.begin(); i != server_map.end(); i++)
 	{
-		len = send(event.ident, client.getRawResponse(), event.data, 0);
-		__D_DISPLAY_SEND(client.getFd(), len, 1, client.getStrResponse());
-		client.setOffset(len);
-		if (len > 0)
-			return set_write_ready(kq, client);
+		if (add_read_event(kq, i->first))
+			return -1;
 	}
-	else
-	{
-		len = send(event.ident, client.getRawResponse(),
-				client.getLenToSend(), 0);
-		__D_DISPLAY_SEND(client.getFd(), len, 2, client.getStrResponse());
-		client.clearResponse();
-	}
-	return len;
+	return 0;
 }
