+++ core/includes/core.hpp
@@ -18,16 +18,19 @@ int		start_server(const std::vector<Server_config*> & server_blocks,
 int		add_read_event(const int kq, const int socket_fd);
 int		set_write_ready(const int kq, Client& client);
 
-//--server_process.cpp--
-int		send_response(const int kq, const struct kevent& event, Client& client);
-int		read_request(const struct kevent& event, Client& client);
+//--server_io_unix.cpp--
+int	send_response(const int kq, const struct kevent& event, Client& client);
+int	read_request(const struct kevent& event, Client& client);
 
-int		accept_new_client(const int kq, const int event_fd,
+//--server_process.cpp--
+int	accept_new_client(const int kq, const int event_fd,
 			std::map<int, Client>& client_map,
 				const std::map<int, std::pair<std::string, int> >& server_map);
 
-int		check_new_connection(const int event_fd,
+int	check_new_connection(const int event_fd,
 			const std::map<int, std::pair<std::string, int> >& server_map);
+int	monitor_network_sockets(const int kq,
+				const std::map<int, std::pair<std::string, int> > & server_map);
 
 int		is_valid_request(Client& client);
