+++ test/drunksite/zozor_site.html
@@ -19,11 +19,9 @@
 
             <nav>
                     <ul>
+                        <li><a href="zozor_site.html">accueil</a></li>
                         <li><a href="test_cgi.html">test cgi</a></li>
-                        <li><a href="/zozor_site.html">accueil</a></li>
-                        <li><a href="#">Blog</a></li>
-                        <li><a href="#">cv</a></li>
-                        <li><a href="#">contact</a></li>
+                        <li><a href="test_post.html">test post</a></li>
                     </ul>
             </nav>
